feat: migrate InstitutionDomainEditor.spec
